Move PostScript to FileMetadata.h
